# lib/api-zod/src/generated/types/quizSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * NIAT SPI Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface QuizSummary {
  attempted: number;
  total: number;
  avgPct: number;
}